Перенёс схему сделанную при помощи библиотеки    Joi    в отдельную папку и сделал её реэкспорт.  Модуль 4     Node.js
Перенёс схему сделанную при помощи библиотеки    Joi    в отдельную папку. Для этого перенёс из файла    contacts.js     (который находится в папке    models    ) в файл     schemas.js    импорт библиотеки   Joi    также перенёс переменные   addSchema    updateFavoriteSchema    и    schemas
   А в файле   index.js    (который находится в папке  joi     ) я сделал реэкспорт схем, сделанных при помощи библиотеки   joi     В связи с выше указанными манипуляциями при импорте схем сделанных при помощи библиотеки   joi    в файл     contacts.js    (который находится в папке    api    ) достаточно было указать только папку (в которой находится файл со  схемами, сделанных при помощи библиотеки   joi    )  Модуль 4     Node.js
