+++ models/contacts.js
@@ -1,7 +1,5 @@
 const { Schema, model } = require("mongoose");
 
-const Joi = require("joi");
-
 const { handleMongooseError } = require("../helpers");
 
 const contactsSchema = Schema(
@@ -28,27 +26,8 @@ const contactsSchema = Schema(
 
 contactsSchema.post("save", handleMongooseError);
 
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-  favorite: Joi.boolean(),
-});
-
-const updateFavoriteSchema = Joi.object({
-  favorite: Joi.boolean()
-    .required()
-    .error((errors) => new Error("missing field favorite")),
-});
-
-const schemas = {
-  addSchema,
-  updateFavoriteSchema,
-};
-
 const ContactMongoose = model("contact", contactsSchema);
 
 module.exports = {
   ContactMongoose,
-  schemas,
 };
